requirements.txt
astroid==2.3.1
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
PyJWT==1.7.1
pylint==2.4.2
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.2
wrapt==1.11.2
